third_party/blink/web_tests/platform/generic/fast/url/mailto-expected.txt

Tests mailto: URL canonicalization

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS canonicalize('mailto:addr1') is 'mailto:addr1'
PASS canonicalize('mailto:addr1@foo.com') is 'mailto:addr1@foo.com'
PASS canonicalize('mailto:addr1 	 ') is 'mailto:addr1'
PASS canonicalize('mailto:addr1?to=jon') is 'mailto:addr1?to=jon'
PASS canonicalize('mailto:addr1,addr2') is 'mailto:addr1,addr2'
PASS canonicalize('mailto:addr1, addr2') is 'mailto:addr1,%20addr2'
PASS canonicalize('mailto:addr1,"addr2"') is 'mailto:addr1,%22addr2%22'
PASS canonicalize('mailto:addr1%2caddr2') is 'mailto:addr1%2caddr2'
PASS canonicalize('mailto:𐌀') is 'mailto:%F0%90%8C%80'
PASS canonicalize('mailto:addr1?') is 'mailto:addr1?'
PASS canonicalize('mailto:í ') == 'mailto:%EF%BF%BD' is false
PASS successfullyParsed is true

TEST COMPLETE
